Add a view controller with binding logic stuck in the editor window.
Look at the action listeners I use. I tried to build it out so if we wanted to
move more logic to the controller then we could, though we would need to set
some methods on the view to get some information like selected shape and stuff.

The private helpers to update fields could be made public.

The assignment spec does seem to suggest a lot of controller control.
But I personally understand the view better this way. It controls its own
buttons and ferries user requests to the model (and does complicated logic
like renaming a shape which is something a user might want to do).

+++ src/cs3500/easyanimator/model/motions/IMotion.java
@@ -136,11 +136,11 @@ public interface IMotion {
    * @return the value at the desired tick.
    */
   static int tween(int val1, int val1T, int val2, int val2T, int tick) {
-    double a = (double) val1;
-    double aT = (double) val1T;
-    double b = (double) val2;
-    double bT = (double) val2T;
-    double t = (double) tick;
+    double a = val1;
+    double aT = val1T;
+    double b = val2;
+    double bT = val2T;
+    double t = tick;
 
     double f1 = (bT - t) / (bT - aT);
     double f2 = a * f1;
